requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
loguru
sqlalchemy>=2.0
alembic
asyncpg
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]